Updated gui for better visibility
A tenths of km digit was added, proper legends put into dials,
increased font size for current and temperature and added
battery icon

+++ ui/gui/include/gui/main_screen/MainPresenter.hpp
@@ -33,7 +33,7 @@ public:
 
     void notifyRpmChange(int value);
     void notifySpeedChange(int value);
-    void notifyDistanceChange(int value);
+    void notifyDistanceChange(double value);
     void notifyBatteryChange(int percentage);
     void notifyMotorTempChange(int temp);
     void notifyControllerTempChange(int temp);
